Found Difference in provided dashboards on import (2)
On the initial import, I found that content is manually needed for this dashboard in order for Grafana to accept the import. For `icinga2-default.json`, and most of the other dashboards "null" is specified and one is able to import the dashboard without any immediate changes. Requesting to update this dashboard file to match the other for consistency.